# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/bootstrap": {
            "post": {
                "description": "One-time endpoint that creates the initial admin user with an HSM wallet. Returns the wallet address for on-chain role assignment. Returns 409 if any user already exists.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Bootstrap admin",
                "parameters": [
                    {
                        "description": "Admin email",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api_adapters_handlers.bootstrapRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "description": "Returns profile information for the authenticated PRIVACY_NODE_OPERATOR.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api_adapters_handlers.meResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/tokens/{address}/permissions": {
            "get": {
                "description": "Returns, for the authenticated user's custody wallet, which Access Manager-gated\nfunctions (mint/burn/...) it can call on the token, plus convenience flags.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tokens"
                ],
                "summary": "Get token permissions for the current user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token contract address (0x...)",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ops-api_cmd_api_core.TokenPermissions"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "cmd_api_adapters_handlers.bootstrapRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
...
        "cmd_api_adapters_handlers.meResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "wallet_address": {
                    "type": "string"
                }
            }
        },
...
        "ops-api_cmd_api_core.TokenFunction": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "friendly name (mint/burn/...), empty if unknown",
                    "type": "string"
                },
                "selector": {
                    "description": "4-byte selector, e.g. \"0x40c10f19\"",
                    "type": "string"
                }
            }
        },
        "ops-api_cmd_api_core.TokenPermissions": {
            "type": "object",
            "properties": {
                "canBurn": {
                    "type": "boolean"
                },
                "canMint": {
                    "type": "boolean"
                },
                "canPause": {
                    "description": "CanPause reports whether this wallet is the stablecoin's `pauser`. Deliberately not\nderived from AM roles: pause()/unpause() are gated on a msg.sender equality check\nagainst the contract's own `pauser` address. False for non-stablecoins, which have no\npause function at all.",
                    "type": "boolean"
                },
                "canSubmitTokenUpdate": {
                    "type": "boolean"
                },
                "contractAddress": {
                    "type": "string"
                },
                "functions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ops-api_cmd_api_core.TokenFunction"
                    }
                },
                "isPaused": {
                    "description": "IsPaused is the ACCESS MANAGER's pause on this managed contract (mirrored from its\nContractPauseUpdated events) — NOT the stablecoin's own `paused` state variable.",
                    "type": "boolean"
                },
                "isTokenPaused": {
                    "description": "IsTokenPaused is the stablecoin's OWN `paused` flag, read live from the contract — the\none that actually halts transfers. Distinct from IsPaused above; false for standards\nwith no pause function.",
                    "type": "boolean"
                },
                "walletAddress": {
                    "type": "string"
                }
            }
        },
...
    }
}